requirements.txt
asgiref==3.2.7
autopep8==1.5.2
Django==3.0.5
django-crispy-forms==1.9.1
django-filter==2.2.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
Markdown==3.2.2
Pillow==7.1.2
pycodestyle==2.5.0
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
